Game Controller: executeNextStep method in gameController corrected so that it only computes the distance once, between all the players in the list before they execute their steps, so when they've all executed their step it can be recalculated with the new list of closestPlayers to PA, so we've created priorityList that holds the current list of calculated closest players.
PriorityAntenna: Setting minDistance to be a huge number, so that when sorting the list that no matter what that our loop, loops for each and every player in the list and calculates their distance accordingly so that all the players get added to the list.
sortedPlayers has been moved into the calcClosestPlayers method so the amount of players stays 6 and so it doesnt add more players than we already have.
Priority Antenna Test:
Now calculates the distance between players to the Priority Antenna and when players have been moved it recalculates their distance and updates the list, so that there's a new closestPlayer.